Isolated base command class, cleaned up roles cmds
Moved the BaseCommand class out of roles into separate file,
renamed as Command in Role controller, cleaned up ctrl array.
